+++ Hunger/Modules/List/DetailListStoryboard.storyboard
@@ -45,9 +45,8 @@
                                         <rect key="frame" x="0.0" y="129.5" width="375" height="493.5"/>
                                         <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                                     </tableView>
-                                    <activityIndicatorView opaque="NO" contentMode="scaleToFill" horizontalHuggingPriority="750" verticalHuggingPriority="750" fixedFrame="YES" hidesWhenStopped="YES" animating="YES" style="medium" translatesAutoresizingMaskIntoConstraints="NO" id="zD6-Ln-dVJ">
-                                        <rect key="frame" x="177" y="301" width="20" height="20"/>
-                                        <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                                    <activityIndicatorView opaque="NO" contentMode="scaleToFill" horizontalHuggingPriority="750" verticalHuggingPriority="750" hidesWhenStopped="YES" animating="YES" style="medium" translatesAutoresizingMaskIntoConstraints="NO" id="zD6-Ln-dVJ">
+                                        <rect key="frame" x="177.5" y="301.5" width="20" height="20"/>
                                     </activityIndicatorView>
                                     <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Encuentra los lugares con buena comida a la medida del gusto y precio justo" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="mhT-wX-ref">
                                         <rect key="frame" x="30" y="23.5" width="321" height="72"/>
@@ -62,8 +61,10 @@
                                 <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                                 <constraints>
                                     <constraint firstItem="mhT-wX-ref" firstAttribute="top" secondItem="sN9-RD-FUb" secondAttribute="top" constant="23.5" id="0BE-GE-2dR"/>
+                                    <constraint firstItem="zD6-Ln-dVJ" firstAttribute="centerY" secondItem="sN9-RD-FUb" secondAttribute="centerY" id="2LV-NC-XYW"/>
                                     <constraint firstItem="mhT-wX-ref" firstAttribute="leading" secondItem="sN9-RD-FUb" secondAttribute="leading" constant="30" id="A2k-4k-bOV"/>
                                     <constraint firstItem="6n3-gD-zSu" firstAttribute="leading" secondItem="sN9-RD-FUb" secondAttribute="leading" id="Hdq-xS-pHy"/>
+                                    <constraint firstItem="zD6-Ln-dVJ" firstAttribute="centerX" secondItem="sN9-RD-FUb" secondAttribute="centerX" id="LWe-SD-1qW"/>
                                     <constraint firstAttribute="trailing" secondItem="6n3-gD-zSu" secondAttribute="trailing" id="Vdd-lA-JxT"/>
                                     <constraint firstItem="6n3-gD-zSu" firstAttribute="top" secondItem="mhT-wX-ref" secondAttribute="bottom" constant="34" id="ZrL-lX-z6C"/>
                                     <constraint firstAttribute="bottom" secondItem="6n3-gD-zSu" secondAttribute="bottom" id="b62-oZ-wto"/>
@@ -76,7 +77,7 @@
                                     <navigationItem id="Nk1-Ih-CAx">
                                         <barButtonItem key="rightBarButtonItem" image="SideBarIcon" id="HTo-Qy-0Yu">
                                             <connections>
-                                                <action selector="SlideButton:" destination="Y6W-OH-hqX" id="I2v-PS-JFS"/>
+                                                <action selector="SlideButton" destination="Y6W-OH-hqX" id="vuT-Vb-sh0"/>
                                             </connections>
                                         </barButtonItem>
                                     </navigationItem>
@@ -91,7 +92,7 @@
                                 <inset key="contentEdgeInsets" minX="10" minY="10" maxX="10" maxY="10"/>
                                 <state key="normal" image="BackButton"/>
                                 <connections>
-                                    <action selector="backButton:" destination="Y6W-OH-hqX" eventType="touchUpInside" id="mQP-BD-iTu"/>
+                                    <action selector="backButton" destination="Y6W-OH-hqX" eventType="touchUpInside" id="Cc1-TM-TQ0"/>
                                 </connections>
                             </button>
                             <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="RESTAURANTES" textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="G9X-Ry-gJz">
@@ -126,8 +127,8 @@
                         <outlet property="descriptionText" destination="mhT-wX-ref" id="2FY-EN-cgO"/>
                         <outlet property="loader" destination="zD6-Ln-dVJ" id="7hp-Jj-ovB"/>
                         <outlet property="messagesTable" destination="6n3-gD-zSu" id="A2q-DI-Pcn"/>
-                        <outlet property="viewLeading" destination="fpq-ud-05j" id="z0V-ZA-fkT"/>
-                        <outlet property="viewTrailing" destination="X3O-PQ-wi4" id="f7L-nk-lkh"/>
+                        <outlet property="viewLeadingConstraint" destination="fpq-ud-05j" id="Sh2-oZ-fNe"/>
+                        <outlet property="viewTrailingConstraint" destination="X3O-PQ-wi4" id="DUY-XD-ek1"/>
                         <segue destination="5kB-Vi-aSK" kind="show" identifier="ShowNotFound" id="8mg-f9-d48"/>
                     </connections>
                 </viewController>
