+++ Hunger/Modules/List/NotFoundViewController.swift
@@ -11,7 +11,7 @@ class NotFoundViewController: UIViewController {
 
 
     @IBAction func reloadButton(_ sender: Any) {
-        navigationController?.dismiss(animated: true)
+        dismiss(animated: true)
     }
     
     override func viewDidLoad() {
