+++ Hunger/Modules/Map/MapStoryboard.storyboard
@@ -31,7 +31,7 @@
                                 <inset key="contentEdgeInsets" minX="10" minY="10" maxX="10" maxY="10"/>
                                 <state key="normal" image="BackButton"/>
                                 <connections>
-                                    <action selector="backButton:" destination="Y6W-OH-hqX" eventType="touchUpInside" id="Ntf-qN-yYM"/>
+                                    <action selector="backButton" destination="Y6W-OH-hqX" eventType="touchUpInside" id="QhC-FC-iKj"/>
                                 </connections>
                             </button>
                             <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="RESTAURANTES" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="kQH-Rs-mFa">
